Servlets can now also use Operation Handles
A servlet which does not produce HTML output can now also use an
operation handle. This handle will then cache all the events (actually,
up to 128 events). The next time a JSP allocates a handle, these
events will be flushed. This is not a very beautiful solution, but for
now it will do.