JavaScript to Set Form Action + Refactoring of Controller Command Names
We now just have one drop down menu with choices what to do with the
selected elements plus one button. JavaScript will take care of setting
for form's action (target URL) based on the selection changes in the
drop down menu. This looks better. Also, we interactively include help
text based on the selection.

The command strings and form element names have been centralized in the
new class ControllerUtils, as well as some of the behavior previously
part of the controller.jsp. This means more data is in a central class
and can easily be changed.