Common HTTP/HTML-`<meta>` Headers
The new class `Header` provides some common header elements to be sent
in `HttpServletResponse`s of any servlet or JSP, as well as in the
`JspWriter`s of the JSPs in form of HTML `<meta>` tags. This keeps the
`<meta>`-tags and the headers in synch and allows me to have several
common values in a central location.