Complete Redesign of Editor Modules Structure
So far, all editor modules we single classes with static methods for
doing their job. This led to a lot of redundant code. There now is a
simple base class for editor modules (such as those for text,
configuration, and evaluation files) and each new editor extends this
base class.

The new class `Page` allows for collecting javascript functions and
putting them out together at the bottom of a form or page. It is
designed such that functions will not be included multiple times, which
will greatly reduce the size of pages where this may happen, in
particular of the evaluation editor.