Use `input type="text"` instead of `input type="number"`
I now use `text` input types instead of `number` input types in the
forms of the editor. The reason is that I cannot get the `number` type
to work reliable: It seems to not be possible to get a value like `1.77`
transmitted or displayed without turning it into something like
`1.6999...`. With `text` input types and appropriate regexp `pattern`s,
I think this problem will not exist anymore.