Value of Dimension Bound now Automatically Updated when Type Changes
If you specify a dimension limit for a dimension of data type `double`
or `float`, you can define it to something like `11.6`. If you then
change the data type to `int`, the boundary cannot remain a floating
point number. Now it will automatically be translated to an integer,
i.e., to `11` in case of the above example. However, beware of
JavaScript's limited support to numerical types: It can only understand
`double` and, hence, not properly represent integers outside the range
of -2^52...2^52 (or something). Thus, during the conversion, some values
may change a bit due to lost precision. Yet, this does not matter,
because if you change your dimension's type from `double` to `long`, you
may want to check and change your dimension's limits as well.