First Step towards Better Icons and Auto-Type Detection
This is the first step towards having better file icons and automatic
file type detection: I moved `FSElement` into the new package
`gui.utils.files` and gave it a more "stand-alone" touch. `FSElement`s
can now be generated with the `fromPath` method.

The question of the icon will be moved into the `EFSElementType` enum,
which will be step-by-step extended to support all the file types that
we can generate with our system.

This will then later be used to open automatically the correct editor
for a given file.