controller.jsp  Uses FunctionRenderer API, BugFix in Viewer Servlet
The `controller.jsp` now uses the FunctionRenderer API to obtain compact
onchange JavaScripts, as well as compact select/de-select all functions.
The new classes developed for that purpose will help me in completing
the `new` button.

The `viewer` servlet had a small bug in its URL redirection code: local
URLs (starting with `#`) would also be redirected, which should not be
done. This is fixed now.