Examples Integrated
The examples (MAX-3SAT, BBOB, TSP Suite) which are part of the
documentation of the project are now also integrated into the GUI: You
can directly download them into your workspace with one click.

We therefore use the new Shell tool of the core code and first download
a batch file from the documentation repository. This batch file is then
piped to the shell.

Having the batch file as part of the documentation will allow me to
modify it if the example should change (for whatever reason).