Hopefully a BugFix for the "Disappearing of New Files"
New files would land in the wrong folder due to their paths being
relativized to their current directory instead of to the root directory.
This means new files landed in the root folder, but not in the current
directory.

This has hopefully been fixed now.

Furthermore, I reduced the number of menu components in the previous
commit to make the UI easier to use.